Integrate main resource into order resource
