+++ hexa_delivery/lib/bloc/main_page_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:hexa_delivery/model/dto.dart';
-import 'package:hexa_delivery/resources/mainpage_provider.dart';
+import 'package:hexa_delivery/resources/order_resource.dart';
 
 class MainPageBloc {
   final StreamController<List<OrderTopDTO>> _orderTopDTOStream =
@@ -9,7 +9,7 @@ class MainPageBloc {
   Stream<List<OrderTopDTO>> get orderTopDTOStream => _orderTopDTOStream.stream;
 
   void requestNewOrderTopDTO() async {
-    List<OrderTopDTO> orderTopDTOList = await MainPageProvider.mainList();
+    List<OrderTopDTO> orderTopDTOList = await OrderResource.getTopOrders();
     _orderTopDTOStream.sink.add(orderTopDTOList);
   }
